Finalize version and keywords
I was on the fence about what version I wanted to start with. I was originally thinking 0.5.0 to allow for a bit of variability in the API until I'm confident of it, but I realized that it's such a small package that there will likely be very few changes anyways, and what I'll be publishing is probably fine.

I also picked a few keywords based on what similar packages had on NPM. There are a few offering what we do, but they're all implementing it in different ways (i.e. not using `deasync`) and were written years ago.